Support resources regardless of capitalization
Previously, I required the EXACT capitalization of the Interface
in order to manage its proxy settings. This commit (along with an oopsie
from a previous commit) allows you to manage proxy information and
capitalize the interface however you want.

The only drawback is `puppet resource` - it CURRENTLY requires you to
use ALL LOWERCASE LETTERS for the Interface name when declaring any
of the proxy resources.  I've not been able to pinpoint WHY `puppet resource`
is special, but there you are...