Make mac_proxy_bypassdomains standalone
Previously, the mac_proxy_bypassdomains provider inherited
from Puppet::Provider::MacProxy - but ONLY for a single method
call.  Unfortunately, when it did that, it ALSO inherited the
prefetching/flushing action…which is unnecessary here. This
commit makes the mac_proxy_bypassdomains standalone, but also
lets it call the method it needs from the mac_proxy provider.